requirements.txt
-f /usr/share/pip-wheels
algoliasearch==1.13.0
algoliasearch-django==1.5.0
asn1crypto==0.23.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
construct==2.5.3
cryptography==2.1.1
dj-database-url==0.4.2
Django==2.0.1
django-addanother==2.0.0
django-debug-toolbar==1.9.1
django-haystack==2.7.dev0
django-heroku==0.2.0
django-pure-pagination==0.3.0
django-taggit==0.22.1
factory-boy==2.9.2
Faker==0.8.5
future==0.16.0
idna==2.6
Markdown==2.6.9
mysqlclient==1.3.12
nose2==0.6.5
olefile==0.44
pefile==2017.9.3
Pillow==4.3.0
psycopg2==2.7.3.2
pycparser==2.18
pyOpenSSL==17.3.0
pysolr==3.6.0
python-dateutil==2.6.1
python-ptrace==0.9.3
pytz==2017.3
reportlab==3.4.0
requests==2.18.4
selenium==3.6.0
six==1.11.0
sqlparse==0.2.4
Unidecode==0.4.21
urllib3==1.22
whitenoise==3.3.1
